refactor:(#8) map 관련 커스텀 훅 및 컴포넌트 분리
