+++ src/components/maps/MapControls.jsx
@@ -1,0 +1,10 @@
+import { MapTypeControl, ZoomControl } from 'react-kakao-maps-sdk';
+
+const MapControls = () => (
+  <>
+    <MapTypeControl position={"TOPRIGHT"} />
+    <ZoomControl position={"RIGHT"} />
+  </>
+);
+
+export default MapControls;
